pyproject.toml
[project]
name = "TemplateProject"
version = "0.1.1"
description = "A Template project for python apps with Jinja2 templating"
authors = [
    { name = "Gemini Code Assist", email = "your@email.com" },
]
requires-python = ">=3.9"
dependencies = [
    "ttkbootstrap>=1.10.1",
    "jinja2>=3.1.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
dev = ["ruff"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
# See https://docs.astral.sh/ruff/rules/
select = ["E", "F", "I"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is used.
fixable = ["A", "B", "C", "D", "E", "F", "I"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 90
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
